+++ packages/sitebuilder/src/hooks/index.ts
@@ -4,7 +4,6 @@ export * from './useFirstTimeConfiguration';
 export * from './useLookAndFeel';
 export * from './useLookAndFeelColors';
 export * from './useLookAndFeelFonts';
-export * from './useSetupCard';
 export * from './useSiteBuilder';
 export * from './useUsernameValidation';
 export * from './useWizard';