+++ packages/sitebuilder/src/setup/footer/index.ts
@@ -1,4 +1,2 @@
-export { default as LookAndFeelFooter } from './LookAndFeelFooter';
-export * from './LookAndFeelFooter';
-export { default as SiteDomainFooter } from './SiteDomainFooter';
-export * from './SiteDomainFooter';
+export { default as FooterColumns } from './FooterColumns';
+export { default as FooterLinks } from './FooterLinks';