+++ packages/wme-ui/src/components/jump-nav/index.ts
@@ -1,0 +1,2 @@
+export * from './jump-nav';
+export { default } from './jump-nav';